make newsfeedTitle a component; resolves PR request
