+++ src/features/twitter-feed/index.jsx
@@ -1,25 +1,22 @@
 import React from 'react';
-import { NewsfeedPanel } from '../../components/NewsfeedPanel';
+import { NewsfeedPanel } from '../../components/NewsfeedPanel/index';
 import { Newsfeed } from './Newsfeed';
+import NewsfeedTitle from '../../components/NewsfeedPanel/NewsfeedTitle';
 import useAppState from 'state/AppStateProvider';
-import { useStyles } from './styles';
 
 function TwitterFeed() {
   const { openLiveStream } = useAppState();
-  const styles = useStyles();
 
-  const title = () => {
-    return (
-      <div className={styles.root}>
-        Twitter Popular Tweets
-        <div>
-          <i class="fab fa-twitter" />
-        </div>
-      </div>
-    );
-  };
   return (
-    <NewsfeedPanel panelTitle={title()} source={'twitter'}>
+    <NewsfeedPanel
+      panelTitle={
+        <NewsfeedTitle
+          title={'Twitter Popular Tweets'}
+          icon={<i className="fab fa-twitter" />}
+        />
+      }
+      source={'twitter'}
+    >
       <Newsfeed openLiveStream />
     </NewsfeedPanel>
   );
